test case 추가

+++ src/main/java/com/circlee7/test/service/impl/ProgramServiceImpl.java
@@ -94,6 +94,13 @@ public class ProgramServiceImpl implements ProgramService {
     @Override
     @Transactional
     public void deleteProgram(String programId) {
-        programRepository.deleteById(programId);
+
+        Optional<Program> program = programRepository.findById(programId);
+        if(program.isPresent()) {
+            programRepository.deleteById(programId);
+        } else {
+            throw new EntityNotFoundException();
+        }
+
     }
 }
